requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
pyyaml
sqlalchemy
psycopg2-binary
pika
python-multipart
pytesseract
pillow
openai
python-docx
PyMuPDF